usuwanie kursow + przycisk na navbar + 3 pola w kursie + zamiana punktow na znizki

+++ app/Http/Controllers/CourseController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Course;
 use App\Registers;
+use Illuminate\Support\Facades\Auth;
+
 
 class CourseController extends Controller
 {
@@ -18,6 +20,10 @@ class CourseController extends Controller
         $course->price=$request->input('price');
         $course->points=$request->input('points');
         $course->slots=$request->input('slots');
+        $course->status=true;
+        $course->toLearn1=$request->input('toLearn1');
+        $course->toLearn2=$request->input('toLearn2');
+        $course->toLearn3=$request->input('toLearn3');
         $course->registered =0;
         $course->discount=0;
 
@@ -28,8 +34,8 @@ class CourseController extends Controller
         $image->move($destinationPath, $imageName);
 
         $course->save();
-        $courses = Course::all();
-        return view('admin.panel')->with('courses', $courses);
+
+        return app('App\Http\Controllers\PagesControler')->panel();
     }
 
 
@@ -61,26 +67,30 @@ class CourseController extends Controller
         $course->points=$request->input('points');
         $course->discount=$request->input('discount');
         $course->slots=$request->input('slots');
+        $course->toLearn1=$request->input('toLearn1');
+        $course->toLearn2=$request->input('toLearn2');
+        $course->toLearn3=$request->input('toLearn3');
+        $course->status=true;
         $course->registered = $course->registered;
 
-        $image = $request->file('image');
-        $imageName = $request->input('title') . '.' . $image->getClientOriginalExtension();
-        $course->img=$imageName;
-        $destinationPath = public_path('/img/courses');
-        $image->move($destinationPath, $imageName);
-
         $course->save();
-        $courses = Course::all();
-        return view('admin.panel')->with('courses', $courses);
+
+        return app('App\Http\Controllers\PagesControler')->panel();
     }
 
     public function delete(Request $request){
       $id = $request->input('id');
       $course =Course::find($id);
       if($course!=null){
-        $course ->delete();
+        if($course->status==true){
+          $course ->status=false;
+          $course->save();
+        }else{
+          $course->delete();
+          return app('App\Http\Controllers\PagesControler')->deleted();
+        }
       }
-      $courses =Course::all();
-      return view('admin.panel')->with('courses', $courses);
+
+      return app('App\Http\Controllers\PagesControler')->panel();
     }
 }
